🎨  AppleLoginManager 로 관리
