+++ Flicker/Global/Utils/AppleLoginManager.swift
@@ -1,79 +1,80 @@
-////
-////  AppleLoginManager.swift
-////  Flicker
-////
-////  Created by 김연호 on 2022/11/17.
-////
 //
-//import UIKit
+//  AppleLoginManager.swift
+//  Flicker
 //
-//import AuthenticationServices
-//import CryptoKit
-//import Firebase
-//import FirebaseAuth
-//import FirebaseFirestore
+//  Created by 김연호 on 2022/11/17.
 //
-//final class AppleLoginManager: NSObject {
-//    static let shared = AppleLoginManager()
-//
-//    fileprivate var currentNonce: String?
-//    
-//
-//    @available(iOS 13, *)
-//    func createAppleIDRequest() -> ASAuthorizationAppleIDRequest {
-//        let appleIDProvider = ASAuthorizationAppleIDProvider()
-//        let request = appleIDProvider.createRequest()
-//        // 애플로그인은 사용자에게서 2가지 정보를 요구함
-//        request.requestedScopes = [.fullName, .email]
-//
-//        let nonce = randomNonceString()
-//        request.nonce = sha256(nonce)
-//        currentNonce = nonce
-//
-//        return request
-//    }
-//
-//    @available(iOS 13, *)
-//    private func sha256(_ input: String) -> String {
-//        let inputData = Data(input.utf8)
-//        let hashedData = SHA256.hash(data: inputData)
-//        let hashString = hashedData.compactMap {
-//            return String(format: "%02x", $0)
-//        }.joined()
-//
-//        return hashString
-//    }
-//
-//    private func randomNonceString(length: Int = 32) -> String {
-//        precondition(length > 0)
-//        let charset: Array<Character> =
-//        Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
-//        var result = ""
-//        var remainingLength = length
-//
-//        while remainingLength > 0 {
-//            let randoms: [UInt8] = (0 ..< 16).map { _ in
-//                var random: UInt8 = 0
-//                let errorCode = SecRandomCopyBytes(kSecRandomDefault, 1, &random)
-//                if errorCode != errSecSuccess {
-//                    fatalError("Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)")
-//                }
-//                return random
-//            }
-//
-//            randoms.forEach { random in
-//                if remainingLength == 0 {
-//                    return
-//                }
-//
-//                if random < charset.count {
-//                    result.append(charset[Int(random)])
-//                    remainingLength -= 1
-//                }
-//            }
-//        }
-//
-//        return result
-//    }
-//
-//}
+
+import UIKit
+
+import AuthenticationServices
+import CryptoKit
+import Firebase
+import FirebaseAuth
+import FirebaseFirestore
+
+final class AppleLoginManager: NSObject {
+    static let shared = AppleLoginManager()
+
+    var currentNonce: String?
+    var currentAppleIdToken: String?
+
+    
+    @available(iOS 13, *)
+    func createAppleIDRequest() -> ASAuthorizationAppleIDRequest {
+        let appleIDProvider = ASAuthorizationAppleIDProvider()
+        let request = appleIDProvider.createRequest()
+        // 애플로그인은 사용자에게서 2가지 정보를 요구함
+        request.requestedScopes = [.fullName, .email]
+
+        let nonce = randomNonceString()
+        request.nonce = sha256(nonce)
+        currentNonce = nonce
+
+        return request
+    }
+
+    @available(iOS 13, *)
+    func sha256(_ input: String) -> String {
+        let inputData = Data(input.utf8)
+        let hashedData = SHA256.hash(data: inputData)
+        let hashString = hashedData.compactMap {
+            return String(format: "%02x", $0)
+        }.joined()
+
+        return hashString
+    }
+
+   func randomNonceString(length: Int = 32) -> String {
+        precondition(length > 0)
+        let charset: Array<Character> =
+        Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
+        var result = ""
+        var remainingLength = length
+
+        while remainingLength > 0 {
+            let randoms: [UInt8] = (0 ..< 16).map { _ in
+                var random: UInt8 = 0
+                let errorCode = SecRandomCopyBytes(kSecRandomDefault, 1, &random)
+                if errorCode != errSecSuccess {
+                    fatalError("Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)")
+                }
+                return random
+            }
+
+            randoms.forEach { random in
+                if remainingLength == 0 {
+                    return
+                }
+
+                if random < charset.count {
+                    result.append(charset[Int(random)])
+                    remainingLength -= 1
+                }
+            }
+        }
+
+        return result
+    }
+
+}
