+++ Flicker/Screens/Login/LoginViewController.swift
@@ -17,11 +17,6 @@ import FirebaseFirestore
 
 final class LogInViewController: BaseViewController {
 
-    static let shared = LogInViewController()
-    
-    var currentNonce: String?
-    var currentAppleIdToken: String?
-
     private lazy var appleLoginButton = ASAuthorizationAppleIDButton(type: .signIn, style: .black).then {
         $0.addTarget(self, action: #selector(handleAuthorizationAppleIDButtonPress), for: .touchUpInside)
     }
@@ -217,34 +212,9 @@ final class LogInViewController: BaseViewController {
         self.navigationController?.pushViewController(viewController, animated: true)
     }
 
-    @available(iOS 13, *)
-    func createAppleIDRequest() -> ASAuthorizationAppleIDRequest {
-        let appleIDProvider = ASAuthorizationAppleIDProvider()
-        let request = appleIDProvider.createRequest()
-        // 애플로그인은 사용자에게서 2가지 정보를 요구함
-        request.requestedScopes = [.fullName, .email]
-
-        let nonce = randomNonceString()
-        request.nonce = sha256(nonce)
-        currentNonce = nonce
-
-        return request
-    }
-
-    @available(iOS 13, *)
-    private func sha256(_ input: String) -> String {
-        let inputData = Data(input.utf8)
-        let hashedData = SHA256.hash(data: inputData)
-        let hashString = hashedData.compactMap {
-            return String(format: "%02x", $0)
-        }.joined()
-
-        return hashString
-    }
-    
     //Apple ID 승인 요청
     @objc private func handleAuthorizationAppleIDButtonPress() {
-        let request = createAppleIDRequest()
+        let request = AppleLoginManager.shared.createAppleIDRequest()
         let authorizationController = ASAuthorizationController(authorizationRequests: [request])
         authorizationController.delegate = self
         authorizationController.presentationContextProvider = self
@@ -290,9 +260,11 @@ extension LogInViewController: ASAuthorizationControllerDelegate, ASAuthorizatio
 
     // 사용자 인증 후 처리
     func authorizationController(controller: ASAuthorizationController, didCompleteWithAuthorization authorization: ASAuthorization) {
+        AppleLoginManager.shared.randomNonceString()
+        
         if let appleIDCredential = authorization.credential as? ASAuthorizationAppleIDCredential {
             // 현재 nonce가 설정되어 있는지 확인
-            guard let nonce = currentNonce else {
+            guard let nonce = AppleLoginManager.shared.currentNonce else {
                 fatalError("Invalid state: A login callback was received, but no login request was sent.")
             }
             // ID 토큰값을 검색
@@ -302,7 +274,7 @@ extension LogInViewController: ASAuthorizationControllerDelegate, ASAuthorizatio
             }
             //문자열로 변환
             guard let idTokenString = String(data: appleIDtoken, encoding: .utf8) else {
-                print("Unable to serialize token string from data: \(currentAppleIdToken.debugDescription)")
+                print("Unable to serialize token string from data: \(appleIDtoken.debugDescription)")
                 return
             }
 
@@ -311,11 +283,14 @@ extension LogInViewController: ASAuthorizationControllerDelegate, ASAuthorizatio
                                                       rawNonce: nonce)
 
             Auth.auth().signIn(with: credential) { (authDataResult, error) in
-                if let user = authDataResult?.user {
-                    //로그인 성공 시
-                    self.goProfile()
-                    self.currentAppleIdToken = idTokenString
-                }
+                guard let user = authDataResult?.user else { return }
+                self.goProfile()
+                AppleLoginManager.shared.currentAppleIdToken = idTokenString
+//                if let user = authDataResult?.user {
+//                    //로그인 성공 시
+//                    self.goProfile()
+//                    self.currentAppleIdToken = idTokenString
+//                }
 
                 if error != nil {
                     print(error?.localizedDescription ?? "error" as Any)
@@ -324,38 +299,4 @@ extension LogInViewController: ASAuthorizationControllerDelegate, ASAuthorizatio
             }
         }
     }
-}
-
-//임의로 생성되는 암호화 토큰이다.
-
-private func randomNonceString(length: Int = 32) -> String {
-    precondition(length > 0)
-    let charset: Array<Character> =
-    Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
-    var result = ""
-    var remainingLength = length
-
-    while remainingLength > 0 {
-        let randoms: [UInt8] = (0 ..< 16).map { _ in
-            var random: UInt8 = 0
-            let errorCode = SecRandomCopyBytes(kSecRandomDefault, 1, &random)
-            if errorCode != errSecSuccess {
-                fatalError("Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)")
-            }
-            return random
-        }
-
-        randoms.forEach { random in
-            if remainingLength == 0 {
-                return
-            }
-
-            if random < charset.count {
-                result.append(charset[Int(random)])
-                remainingLength -= 1
-            }
-        }
-    }
-
-    return result
 }
