+++ Flicker/Screens/ProfileSetting/InputPassword/InputPasswordViewController.swift
@@ -111,8 +111,8 @@ final class InputPasswordViewController: BaseViewController {
         //TODO: -
         // 애플재인증과 이메일 재인증을 분기처리하기 위해 서버에 이메일,애플로그인 구별을 하던
         // 코어데이터를 활용해서 하든 이메일로그인 인지 애플로그인 인지 식별을 해야함
-        reAuthUser()
-        //        appleLoginReAuthUser()
+//        reAuthUser()
+        appleLoginReAuthUser()
     }
     //사용자 재인증 함수, 이메일은 파베에서 가져오고 비밀번호는 사용자가 입력 한 후, 비밀번호가 틀릴 시 재입력요구, 성공하면 ProfileSettingView로 넘어감
     private func reAuthUser() {
@@ -136,8 +136,8 @@ final class InputPasswordViewController: BaseViewController {
         // Initialize a fresh Apple credential with Firebase.
         let credential = OAuthProvider.credential(
             withProviderID: "apple.com",
-            idToken: LogInViewController.shared.currentAppleIdToken ?? "",
-            rawNonce: LogInViewController.shared.currentNonce
+            idToken: AppleLoginManager.shared.currentAppleIdToken ?? "",
+            rawNonce: AppleLoginManager.shared.currentNonce
         )
         // Reauthenticate current Apple user with fresh Apple credential.
         Auth.auth().currentUser?.reauthenticate(with: credential) { (authResult, error) in
